webhook-deploys/pyproject.toml
[project]
name = "webhook-deploys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.0",
    "pydantic>=2.11.7",
    "uvicorn>=0.35.0",
    "apscheduler>=3.10.4",
    "GitPython>=3.1.43",
    "asyncio>=3.4.3",
]
